refactor to reuse components between reviews
